+++ CHANGELOG.md
@@ -19,6 +19,7 @@ _**For better traceability add the corresponding GitHub issue number in each cha
 - Marked createAccessPolicy requests with deprecation mark.
 - remove edc namespace from EdcContractAgreementNegotiationResponse and
   EdcContractAgreementsResponse eclipse-tractusx/traceability-foss#963
+- added missing @context values in edc asset creation eclipse-tractusx/traceability-foss#978
 
 ## Added